+++ crates/stackable-versioned-macros/src/codegen/container/mod.rs
@@ -8,7 +8,7 @@ use syn::{Attribute, Ident, ItemEnum, ItemStruct, Path, Visibility, parse_quote}
 use crate::{
     attrs::{
         container::StandaloneContainerAttributes,
-        k8s::{KubernetesArguments, KubernetesCrateArguments},
+        k8s::{KubernetesArguments, KubernetesCrateArguments, RawKubernetesOptions},
     },
     codegen::{
         VersionDefinition,
@@ -21,7 +21,7 @@ mod r#enum;
 mod r#struct;
 
 /// Contains common container data shared between structs and enums.
-pub(crate) struct CommonContainerData {
+pub struct CommonContainerData {
     /// Original attributes placed on the container, like `#[derive()]` or `#[cfg()]`.
     pub(crate) original_attributes: Vec<Attribute>,
 
@@ -37,7 +37,7 @@ pub(crate) struct CommonContainerData {
 /// Abstracting away with kind of container is generated makes it possible to create a list of
 /// containers when the macro is used on modules. This enum provides functions to generate code
 /// which then internally call the appropriate function based on the variant.
-pub(crate) enum Container {
+pub enum Container {
     Struct(Struct),
     Enum(Enum),
 }
@@ -51,16 +51,37 @@ impl Container {
         }
     }
 
-    /// Generates the container `From<Version> for NextVersion` implementation.
-    pub(crate) fn generate_from_impl(
+    /// Generates the `From<Version> for NextVersion` implementation for the container.
+    pub fn generate_upgrade_from_impl(
         &self,
         version: &VersionDefinition,
         next_version: Option<&VersionDefinition>,
         add_attributes: bool,
     ) -> Option<TokenStream> {
         match self {
-            Container::Struct(s) => s.generate_from_impl(version, next_version, add_attributes),
-            Container::Enum(e) => e.generate_from_impl(version, next_version, add_attributes),
+            Container::Struct(s) => {
+                s.generate_upgrade_from_impl(version, next_version, add_attributes)
+            }
+            Container::Enum(e) => {
+                e.generate_upgrade_from_impl(version, next_version, add_attributes)
+            }
+        }
+    }
+
+    /// Generates the `From<NextVersion> for Version` implementation for the container.
+    pub fn generate_downgrade_from_impl(
+        &self,
+        version: &VersionDefinition,
+        next_version: Option<&VersionDefinition>,
+        add_attributes: bool,
+    ) -> Option<TokenStream> {
+        match self {
+            Container::Struct(s) => {
+                s.generate_downgrade_from_impl(version, next_version, add_attributes)
+            }
+            Container::Enum(e) => {
+                e.generate_downgrade_from_impl(version, next_version, add_attributes)
+            }
         }
     }
 
@@ -74,7 +95,7 @@ impl Container {
     ///
     /// This function only returns `Some` if it is a struct. Enums cannot be used to define
     /// Kubernetes custom resources.
-    pub(crate) fn generate_kubernetes_item(
+    pub fn generate_kubernetes_item(
         &self,
         version: &VersionDefinition,
     ) -> Option<(IdentString, String, TokenStream)> {
@@ -88,7 +109,7 @@ impl Container {
     ///
     /// This function only returns `Some` if it is a struct. Enums cannot be used to define
     /// Kubernetes custom resources.
-    pub(crate) fn generate_kubernetes_merge_crds(
+    pub fn generate_kubernetes_merge_crds(
         &self,
         enum_variant_idents: &[IdentString],
         enum_variant_strings: &[String],
@@ -108,7 +129,14 @@ impl Container {
         }
     }
 
-    pub(crate) fn get_original_ident(&self) -> &Ident {
+    pub fn generate_kubernetes_status_struct(&self) -> Option<TokenStream> {
+        match self {
+            Container::Struct(s) => s.generate_kubernetes_status_struct(),
+            Container::Enum(_) => None,
+        }
+    }
+
+    pub fn get_original_ident(&self) -> &Ident {
         match &self {
             Container::Struct(s) => s.common.idents.original.as_ident(),
             Container::Enum(e) => e.common.idents.original.as_ident(),
@@ -180,9 +208,17 @@ impl StandaloneContainer {
 
             // NOTE (@Techassi): Using '.copied()' here does not copy or clone the data, but instead
             // removes one level of indirection of the double reference '&&'.
-            let from_impl =
+            let next_version = versions.peek().copied();
+
+            // Generate the From impl for upgrading the CRD.
+            let upgrade_from_impl =
                 self.container
-                    .generate_from_impl(version, versions.peek().copied(), false);
+                    .generate_upgrade_from_impl(version, next_version, false);
+
+            // Generate the From impl for downgrading the CRD.
+            let downgrade_from_impl =
+                self.container
+                    .generate_downgrade_from_impl(version, next_version, false);
 
             // Add the #[deprecated] attribute when the version is marked as deprecated.
             let deprecated_attribute = version
@@ -210,7 +246,8 @@ impl StandaloneContainer {
                     #container_definition
                 }
 
-                #from_impl
+                #upgrade_from_impl
+                #downgrade_from_impl
             });
         }
 
@@ -222,6 +259,8 @@ impl StandaloneContainer {
             false,
         ));
 
+        tokens.extend(self.container.generate_kubernetes_status_struct());
+
         tokens
     }
 }
@@ -231,13 +270,13 @@ impl StandaloneContainer {
 pub(crate) struct ContainerIdents {
     /// The ident used in the context of Kubernetes specific code. This ident
     /// removes the 'Spec' suffix present in the definition container.
-    pub(crate) kubernetes: IdentString,
+    pub kubernetes: IdentString,
 
     /// The original ident, or name, of the versioned container.
-    pub(crate) original: IdentString,
+    pub original: IdentString,
 
     /// The ident used in the [`From`] impl.
-    pub(crate) from: IdentString,
+    pub from: IdentString,
 }
 
 impl ContainerIdents {
@@ -261,32 +300,35 @@ impl ContainerIdents {
 }
 
 #[derive(Debug)]
-pub(crate) struct ContainerOptions {
-    pub(crate) kubernetes_options: Option<KubernetesOptions>,
-    pub(crate) skip_from: bool,
+pub struct ContainerOptions {
+    pub kubernetes_options: Option<KubernetesOptions>,
+    pub skip_from: bool,
 }
 
+// TODO (@Techassi): Get rid of this whole mess. There should be an elegant way of using the
+// attributes directly (with all defaults set and validation done).
 #[derive(Debug)]
-pub(crate) struct KubernetesOptions {
-    pub(crate) group: String,
-    pub(crate) kind: Option<String>,
-    pub(crate) singular: Option<String>,
-    pub(crate) plural: Option<String>,
-    pub(crate) namespaced: bool,
+pub struct KubernetesOptions {
+    pub group: String,
+    pub kind: Option<String>,
+    pub singular: Option<String>,
+    pub plural: Option<String>,
+    pub namespaced: bool,
     // root
-    pub(crate) crates: KubernetesCrateOptions,
-    pub(crate) status: Option<Path>,
+    pub crates: KubernetesCrateOptions,
+    pub status: Option<Path>,
     // derive
     // schema
     // scale
     // printcolumn
-    pub(crate) shortnames: Vec<String>,
+    pub shortnames: Vec<String>,
     // category
     // selectable
     // doc
     // annotation
     // label
-    pub(crate) skip_merged_crd: bool,
+    pub skip_merged_crd: bool,
+    pub config_options: KubernetesConfigOptions,
 }
 
 impl From<KubernetesArguments> for KubernetesOptions {
@@ -303,22 +345,27 @@ impl From<KubernetesArguments> for KubernetesOptions {
             status: args.status,
             shortnames: args.shortnames,
             skip_merged_crd: args.skip.is_some_and(|s| s.merged_crd.is_present()),
+            config_options: args.options.into(),
         }
     }
 }
 
 #[derive(Debug)]
-pub(crate) struct KubernetesCrateOptions {
-    pub(crate) kube_core: Override<Path>,
-    pub(crate) k8s_openapi: Override<Path>,
-    pub(crate) schemars: Override<Path>,
-    pub(crate) serde: Override<Path>,
-    pub(crate) serde_json: Override<Path>,
+pub struct KubernetesCrateOptions {
+    pub kube_client: Override<Path>,
+    pub kube_core: Override<Path>,
+    pub k8s_openapi: Override<Path>,
+    pub schemars: Override<Path>,
+    pub serde: Override<Path>,
+    pub serde_json: Override<Path>,
+    pub versioned: Override<Path>,
 }
 
 impl Default for KubernetesCrateOptions {
     fn default() -> Self {
         Self {
+            versioned: Override::Default(parse_quote! { ::stackable_versioned }),
+            kube_client: Override::Default(parse_quote! { ::kube::client }),
             k8s_openapi: Override::Default(parse_quote! { ::k8s_openapi }),
             serde_json: Override::Default(parse_quote! { ::serde_json }),
             kube_core: Override::Default(parse_quote! { ::kube::core }),
@@ -344,12 +391,20 @@ impl From<KubernetesCrateArguments> for KubernetesCrateOptions {
             crate_options.kube_core = Override::Overridden(kube_core);
         }
 
+        if let Some(kube_client) = args.kube_client {
+            crate_options.kube_client = Override::Overridden(kube_client);
+        }
+
         if let Some(schemars) = args.schemars {
             crate_options.schemars = Override::Overridden(schemars);
         }
 
         if let Some(serde) = args.serde {
             crate_options.serde = Override::Overridden(serde);
+        }
+
+        if let Some(versioned) = args.versioned {
+            crate_options.versioned = Override::Overridden(versioned);
         }
 
         crate_options
@@ -361,11 +416,13 @@ impl ToTokens for KubernetesCrateOptions {
         let mut crate_overrides = TokenStream::new();
 
         let KubernetesCrateOptions {
+            kube_client: _,
             k8s_openapi,
             serde_json,
             kube_core,
             schemars,
             serde,
+            ..
         } = self;
 
         if let Override::Overridden(k8s_openapi) = k8s_openapi {
@@ -396,7 +453,7 @@ impl ToTokens for KubernetesCrateOptions {
 
 /// Wraps a value to indicate whether it is original or has been overridden.
 #[derive(Debug)]
-pub(crate) enum Override<T> {
+pub enum Override<T> {
     Default(T),
     Overridden(T),
 }
@@ -408,6 +465,19 @@ impl<T> Deref for Override<T> {
         match &self {
             Override::Default(inner) => inner,
             Override::Overridden(inner) => inner,
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct KubernetesConfigOptions {
+    experimental_conversion_tracking: bool,
+}
+
+impl From<RawKubernetesOptions> for KubernetesConfigOptions {
+    fn from(options: RawKubernetesOptions) -> Self {
+        Self {
+            experimental_conversion_tracking: options.experimental_conversion_tracking.is_present(),
         }
     }
 }